# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UXUS — Agent Services",
    "version": "1.2.0",
    "description": "Pay-per-call primitives for AI agents. LLM inference, web scraping, structured extraction, embeddings, live search, and EVM wallet/token risk scoring. Paid in USDC on Base via the x402 protocol (v2). No accounts, no API keys.",
    "contact": {
      "name": "UXUS",
      "url": "https://uxus.finance"
    }
  },
  "servers": [
    {
      "url": "https://uxus.finance"
    }
  ],
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "x402-payment",
        "description": "x402 payment, protocol v2. An unpaid request returns 402 with the payment requirement base64-encoded in the PAYMENT-REQUIRED response header and mirrored as JSON in the body. Network: Base (eip155:8453). Asset: USDC (0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913). Pay to 0xe0ed7a30589fec49e98f2085c7162b90fdbb83de. Sign with an x402-capable client and retry with the X-PAYMENT header."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "bad_request"
          },
          "detail": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "PaymentRequired": {
        "type": "object",
        "description": "x402 v2 payment requirement. Also delivered base64-encoded in the PAYMENT-REQUIRED response header.",
        "properties": {
          "x402Version": {
            "type": "integer",
            "example": 2
          },
          "error": {
            "type": "string",
            "example": "Payment required"
          },
          "resource": {
            "type": "object"
          },
          "accepts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "scheme": {
                  "type": "string",
                  "example": "exact"
                },
                "network": {
                  "type": "string",
                  "example": "eip155:8453"
                },
                "amount": {
                  "type": "string",
                  "example": "10000"
                },
                "asset": {
                  "type": "string",
                  "example": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
                },
                "payTo": {
                  "type": "string",
                  "example": "0xe0ed7a30589fec49e98f2085c7162b90fdbb83de"
                },
                "maxTimeoutSeconds": {
                  "type": "integer",
                  "example": 300
                },
                "extra": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "Reason": {
        "type": "object",
        "required": [
          "code",
          "severity",
          "detail",
          "source"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "SANCTIONED"
          },
          "severity": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10
          },
          "detail": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "example": "ofac"
          }
        }
      },
      "RiskFree": {
        "type": "object",
        "required": [
          "address",
          "risk_score",
          "risk_level",
          "verdict",
          "flags"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string",
            "enum": [
              "ethereum",
              "base"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "wallet",
              "token"
            ]
          },
          "risk_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PROCEED",
              "CAUTION",
              "BLOCK"
            ]
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "SANCTIONED"
            ]
          },
          "degraded": {
            "type": "boolean",
            "description": "true = an input needed for the verdict was unavailable (wallet-history explorer failed, or the token-security provider returned nothing) AND there was no sanctions/scam-list hit. verdict is then CAUTION (\"not fully assessed\"), never PROCEED, and carries a HISTORY_UNAVAILABLE or TOKEN_CHECK_UNAVAILABLE flag. Retry before acting."
          },
          "tier": {
            "type": "string",
            "example": "free"
          },
          "upgrade": {
            "type": "string"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time"
          },
          "disclaimer": {
            "type": "string",
            "description": "Ships inside the response: developer-grade screening from public data, not legal/financial/compliance advice. See lib/disclaimer.ts."
          }
        }
      },
      "RiskFull": {
        "type": "object",
        "required": [
          "address",
          "risk_score",
          "risk_level",
          "verdict",
          "flags",
          "reasons",
          "signals",
          "sources"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string",
            "enum": [
              "ethereum",
              "base"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "wallet",
              "token"
            ]
          },
          "risk_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PROCEED",
              "CAUTION",
              "BLOCK"
            ]
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reason"
            }
          },
          "signals": {
            "type": "object",
            "description": "Token signals (tax, mint, proxy, owner flags) or wallet signals (age, tx counts)."
          },
          "sources": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "ofac",
              "scamsniffer"
            ]
          },
          "degraded": {
            "type": "boolean",
            "description": "true = a verdict input was unavailable and there was no list hit; verdict is CAUTION, not PROCEED. See HISTORY_UNAVAILABLE / TOKEN_CHECK_UNAVAILABLE in flags/reasons."
          },
          "checked_at": {
            "type": "string",
            "format": "date-time"
          },
          "cache_ttl": {
            "type": "integer"
          },
          "disclaimer": {
            "type": "string",
            "description": "Ships inside the response: developer-grade screening from public data, not legal/financial/compliance advice. See lib/disclaimer.ts."
          }
        }
      },
...
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. The x402 v2 challenge is in the PAYMENT-REQUIRED header (base64) and mirrored in the body.",
        "headers": {
          "PAYMENT-REQUIRED": {
            "description": "Base64-encoded x402 v2 payment requirement JSON.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PaymentRequired"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid or missing parameters.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/risk": {
      "get": {
        "operationId": "walletRiskFree",
        "summary": "FREE risk demo",
        "description": "Returns risk_score (0-100), risk_level, PROCEED/CAUTION/BLOCK verdict, and flags for any EVM wallet or token. No reasons, signals, or sources.",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "0x0330070FD38Ec3bB94F58FA55D40368271E9e54A"
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ethereum",
                "base"
              ],
              "default": "base"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "wallet",
                "token"
              ],
              "default": "wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Risk summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskFree"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/risk/bazaar": {
      "get": {
        "operationId": "walletRiskBazaar",
        "summary": "Wallet/token risk score via x402 Bazaar ($0.01 USDC on Base)",
        "description": "The same full report as /api/risk/pro, served through the x402 Bazaar resource-server flow (eip155:8453). Returns the free summary when no payment is presented.",
        "security": [
          {
            "x402": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ethereum",
                "base"
              ],
              "default": "base"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "wallet",
                "token"
              ],
              "default": "wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Risk score",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskFull"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          }
        }
      }
    },
...
  }
}